A little cleanup of tests; moving towards using again
