+++ tests/common/mod.rs
@@ -20,36 +20,6 @@ use self::proton_cli::utils;
 use self::proton_cli::project_types::{Project, User};
 
 
-/// Creates a key file at the given location
-/// Returns the path to the key file
-pub fn make_key_file<P: AsRef<Path>>(
-    root_dir: P,
-    file_name: &str,
-    test_key: rsa_keys::TestKey
-) -> PathBuf {
-
-    let file_content = rsa_keys::get_test_key(test_key);
-    write_to_file(root_dir, file_name, &file_content)
-}
-
-/// Write a string to a file, replacing it's current contents if it exists
-/// Returns the path to the file
-fn write_to_file<P: AsRef<Path>>(
-    root_dir: P,
-    file_name: &str,
-    content: &str
-) -> PathBuf {
-
-    let mut file_path = PathBuf::new();
-    file_path.push(root_dir);
-    file_path.push(file_name);
-
-    File::create(&file_path)
-        .and_then(|mut file| write!(file, "{}\n", content))
-        .expect("Error writing to file");
-
-    file_path
-}
 
 /// Returns the path to a music file in /.../cli/tests/music/
 pub fn get_music_file_path(file_name: &str) -> PathBuf {
@@ -63,15 +33,15 @@ pub fn get_music_file_path(file_name: &str) -> PathBuf {
 /// Check if the public key at the given path exists and contains key_content,
 /// and check to see that the user is in the project at the current directory's protonfile
 pub fn assert_user_added<P: AsRef<Path>>(public_key_path: P, name: &str) {
-    let pub_key_contents = utils::file_as_string(public_key_path)
-        .expect("Error reading public key file");
+    // let pub_key_contents = utils::file_as_string(public_key_path)
+    //     .expect("Error reading public key file");
 
-    let project: Project = utils::read_protonfile(None::<P>)
-        .expect("Error reading project");
+    // let project: Project = utils::read_protonfile(None::<P>)
+    //     .expect("Error reading project");
         
-    let u = User::new(name, &pub_key_contents).expect("Error making user for comparison");
-    assert_eq!(project.user_exists(&u), true);
-    assert_eq!(u.permissions.len(), 0);
+    // let u = User::new(name, &pub_key_contents).expect("Error making user for comparison");
+    // assert_eq!(project.user_exists(&u), true);
+    // assert_eq!(u.permissions.len(), 0);
 }
 
 /// Check that changes were actually committed to the repository
