+++ tests_old/rsa_keys/bad_priv_key.pem
@@ -1,0 +1,28 @@
+-----BEGIN R
+INVALID DATA ADDED!!!
+MIIEpAIBAAKCAQEAw9L3r5AgkzsGPizUBMgHLAYlFr5ebMkRRIDLrlVaglv0G1gX
+YgsvWfOKN27CUOegAWW9cpOTBeUHSX9BX4/iq+3FnSGA/G8YMxKnomokLGwS3Wmo
++sXlQSPqJcZjoHhnHHZg9R2DMj09fmHqocu8Zz000To8LK8iFhws+f8Knx3uShsz
+DlR15jcXvxh9YIgCtMDHx3xLasLrTIop9MTNrj3OcKy3rP+o8CoO9iLJdPj95jgr
+L4mnrbKuW0RmdGqqiTVmKlyCQxwXONbs6QH4Cg3CL0W57/4Y5+hU9B0YQ7wu87N/
+M9a8nHNsSDcn4YbdDHajeOMzHtApXZilZ9idVQIDAQABAoIBAG3hZUZIE19KEXi1
+4spEC3IyiIE4lPdp45ySwLe0ATOAsXredVB4gjtxIq90IOkQ74GZOzKdiDQYjXw6
+Qc6DiGd85hQDhsMe+289Gn9SG7YpdnRzI3SDst5Ybrx/k8uphAI+kq03wDRilKzj
++KqNDczIVWjJypYwN2o8qSgmailZvrQZaqc5AtcVeW7WNj19+GhG91EhWDuk5wOa
+ADHMqGLgbtNQTJh/Ur/iFv84eXQKq51jQVyD/fNEOIMy65dOm5DfTkrIdr0FLoZ2
+MXrR91UvSmzfbGrMMXosw+L+L7xoO5+ihhAd5clgRf+apCz9EhUUBTPXqjBgVh7t
+0x1OccECgYEA/3h0J4oNPUvAOAVKBxT8Ol++wjIKBoQh12AEYF15RsrUe6xLSKFI
+Y3YKzqstK9rIytWPcqmyy1eESQ/4vkAiN+Ks+J8nN9YuFEx87fexngRqb0U1vYQP
+qQwK9NjSDyTxEspSgAfPK0BGFDq6Iab03djRbCEXdj1Md8MWE4owLV0CgYEAxDrd
+7HpowsmMob0tpTQOgVLarbc/G2PQ20g5vfsPeepabYrmLNZAqukmtSqQh+SXXtME
+M+xCsK393K/WK5fGRtpOWPODXBHDZYqfMWfwz9DEukNA41vfwqRIbZJcbySdBmo+
+m/6u2hWWc78Dus+l1+ycMTQz80qPFEe2B0iWuFkCgYEA4FYYOfhis+bnPsp2PU3S
+SDh3vPRQFXTGeujYLv8mVmIawReFbJwkgyLPTrjGi/ItzOBIfXpDhYiuo7IgzoBn
+DAhDuHeZ6tBr2mtocVfHY2mhRWDTVj63uczXejdEDwRzpfqBA+jXpbVlVLJVbOO9
+t2qETJ2qiqw9kBWCN1psAAkCgYBj6PUFam3cf3IzB3cuHgsknWLp/9sJvxCgFW4Q
+5LtEf6X7EDWRrqTOYFds9ncQObs4dIg3R5vNCUmwC3coSuaRPKWXdKT1Q7YRBTt4
+lEtwBAOmMoffOu/60lpuL48Qp7urgHzQQCQWtdccp6zmeMMf5UDBbKrmvJeM49Yy
+bysOQQKBgQC9jd6SzKJARDRH79EnmAhGS/rVCl2iZViQ2DK23ElhpkcHj95cKo9u
+K/eaFXWNOcUfBL6FR0YqtQ6fVhPQNjeo805RKge9uo9IAIth4ZTWdr4mxMzzwZ5X
+4cKJdH+pIfGX5ObC+ggEuJdYml3vbfgnldCD4kvhG5GZ1yIBsaDlfQ==
+-----END RSA PRIVATE KEY-----
