+++ tests_old/rsa_keys/bad_pub_key.pub
@@ -1,0 +1,1 @@
+ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDD0vevkCCTOwY+LNQEyAcsBiUWvl5syRFEgMuuVVqCW/QbWBdiCy9Z84o3bsJQ56ABZb1yk5MF5QdJf0Ffj+Kr7cWdIYD8bxgzEqeiaiQsbBLdaaj6xeVBI+olxmOgeGccdmD1HYMyPT1+Yeqhy7xnPTTROjwsryIWHCz5/wqfHe5KGzMOVHXmNxe/GH1giAK0wMfHfEtqwutMiin0xM2uPc5srLes/6jwKg72Isl0+P3mOCsviaetsq5bRGZ0aqqJNWYqXIJDHBc41ucpAfgKDcIvRbnv/hjn6FT0HRhDvC7zs38z1rycc2xINyfhht0MdqN44zMe0CldmKVn2J1V foo@bar
