+++ tests_old/rsa_keys/good_key.pub
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAx53amIap4iq412IBgewF
+R9bjFdRKao9JWoxcNc4CyzjIhRi3ewlsUL6jZGvPJ31O9Ra7EBUKSc7GBeN8tHdL
+fcFUqc9p24YrR85SyVqwNHyXaGWlp5ThLC4I4kN0MtSk02nRpEGbtDwVYS5SK4uj
+Pmlzc1s4LtYIXDf34Hg4bVB8G6h08BZJI1VJz1NANRO/bXvdEFtJKifx4d9vQMZ7
+lDz/IWvSZxe0ff6mN4MVZb5HGwRAbbiqkGfwVTqAHt1zAm3gZ2OmaWKgZrpVfqRH
+Yfx5bWqxNm5AmnE9j1EnEGwQRJ7tznFqv/9SFNGtzrIojKnbC5T2rYuHtz0JerDF
+0QIDAQAB
+-----END PUBLIC KEY-----
