+++ tests_old/rsa_keys/good_key2.pub
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzZmS0oVZ/pR6q1nvOSxB
+B20SsDAIQD2dg6zsjGXDHUDqH4XDPN5db5J8s8baCwCxMgQEniLd4fjCcU4pAZck
+n1B/OE0wWwivOjGsDEP4xAKTLYHEaPHnwudx6e83TG8Kyg8XXMQX9KzgOuWDiYFM
+Q3S/t1DcTO3P1CEez/Iz6Jr2GySz0S07TVRbz2/emt3UUXL2JSxEXwnDY2I+O/52
+KEVeogjXGfy/xcCvdeGhiMecQlPJXtzVy58hzCMWnezcxGUPHWoGs3ZNcu/Us/MU
+8LTe9xiXcafiwve+VZdXIae+ATrhLoa7IYnaNcScEJqhX6cEzki9eAXHg3vhqZ5w
+OwIDAQAB
+-----END PUBLIC KEY-----
