+++ tests_old/rsa_keys/root.pub
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAx2ZLCuIC9D8xQTM79/qX
++inqnhH5Mpdkth0AqZcj9kkIMHgFP0jJtJDBoGMVH29krkzBbrPnSQQ1wYYkmxnm
+R4eHKrdVwANYOHfCyRbNNrI8MOd79hODYBi7chB+G50XN6sOn/2V3Cd0HwNysSoC
+PXoAHirWSQyHGS3QQZrREluKrB5OLuwFTx5RPY6j3r+zy6WNXuoRU5HLCFMx6uSr
+Svn5V67jQqmlotq9UbtdypVsRNAvC1YCAAYpj4yKJwKBYnQTChBbdRinhvkfATxl
+ggg7HkNoXacCrqmHt+E9fruc4xw3geen+F+YsMw1teK9X1uZGdyIlUmHUrOxZtJm
+5wIDAQAB
+-----END PUBLIC KEY-----
